requirements.txt
Django==1.11.29
django-cleanup==4.0.0
django-crispy-forms==1.7.2
Pillow==6.2.2
django-simple-captcha==0.5.12
